DOC: Export faulpolygons in separate script
